Code refactor, UI improvements
