+++ package.json
@@ -72,9 +72,9 @@
     "titlebar": "^1.4.0",
     "uniqid": "^5.0.3",
     "vue": "^2.5.22",
-    "vue-template-compiler": "^2.5.22",
     "vue-electron": "^1.0.6",
     "vue-router": "^3.0.1",
+    "vue-template-compiler": "^2.5.22",
     "vuetify": "^1.4.3",
     "vuex": "^3.1.0",
     "winston": "^3.1.0",
