--- src/main/database/helper.js
@@ -1,9 +1,0 @@
-
-const fsExtra = require('fs-extra')
-// const path = require('path')
-
-export default {
-  ensureDatabaseDir: function (config) {
-    fsExtra.ensureDirSync(config.path)
-  }
-}
