+++ src/main/database/index.js
@@ -1,16 +1,25 @@
+/**
+ * SQL database wrapper
+ */
+
 'use strict'
 import sql from './sql'
-import helper from './helper'
-import commands from './commands'
 
+/**
+ * Sql database wrapper class
+ */
 const NodeSqlite3 = function (props) {
   this.config = props
   this.db = null
 }
 
+/**
+ * Initialize SQL database
+ * Creates/ensures necessary tables for SQL database
+ */
 NodeSqlite3.prototype.init = function (callback) {
   const context = this
-  helper.ensureDatabaseDir(this.config)
+  sql.ensureDatabaseDir(this.config)
   sql.ensureDb(context.config, function (err, db) {
     if (err) {
       console.log(err)
@@ -35,10 +44,13 @@ NodeSqlite3.prototype.init = function (callback) {
   })
 }
 
-NodeSqlite3.prototype.readFullTable = commands.readFullTable
-NodeSqlite3.prototype.readTableEntryID = commands.readTableByID
-NodeSqlite3.prototype.writeTable = commands.writeTable
-NodeSqlite3.prototype.updateTableByID = commands.updateTableByID
-NodeSqlite3.prototype.deleteTableById = commands.deleteTableById
+/**
+ * SQL Database commands API
+ */
+NodeSqlite3.prototype.readFullTable = sql.readFullTable
+NodeSqlite3.prototype.readTableEntryID = sql.readTableByID
+NodeSqlite3.prototype.writeTable = sql.writeTable
+NodeSqlite3.prototype.updateTableByID = sql.updateTableByID
+NodeSqlite3.prototype.deleteTableById = sql.deleteTableById
 
 export default NodeSqlite3
