+++ src/renderer/App.vue
@@ -49,21 +49,37 @@
 </script>
 
 <style>
-  /* CSS */
-  #app { font-family: 'Roboto'; background: transparent; user-select: none; }
+  #app {
+    font-family: 'Roboto';
+    background: transparent;
+    user-select: none;
+  }
+
+  .fade-enter-active, .fade-leave-active {
+    transition: opacity .5s;
+  }
+
+  .fade-enter, .fade-leave-to /* .fade-leave-active below version 2.1.8 */ {
+    opacity: 0;
+  }
+
   .drag-safe {
     -webkit-app-region: no-drag;
   }
+
   ::-webkit-scrollbar {
     width: 0px;
     background: transparent;
   }
+
   ::-webkit-scrollbar-thumb {
     background: transparent;
   }
+
   .light-text {
     font-family: 'Roboto-Light'
   }
+
   #application-bar {
     position: fixed;
     top: 1px;
@@ -78,24 +94,30 @@
     border-bottom-width: 1px;
     z-index: 500;
   }
+
   #application-bar[small] {
     height: 38px;
   }
+
   #application-bar[white] {
     background: #FFFFFF;
   }
+
   #application-bar > .space-left {
     margin-left: 70px;
   }
+
   #application-bar .btn {
     color: #383838;
   }
+
   #application-bar .app-title {
     font-family: 'Roboto-Medium';
     font-size: 14px;
     padding: 0 10px;
     margin: 10px 0;
   }
+
   #sidebar {
     position: fixed;
     top: 61px;
@@ -107,9 +129,11 @@
     border-right-color: rgba(229, 229, 229, 0.596);
     border-right-width: 1px;
   }
+
   #sidebar[small] {
     top: 39px;
   }
+
   #main-content {
     position: fixed;
     top: 60px;
@@ -118,9 +142,11 @@
     right: 0;
     background: white;
   }
+
   #main-content[small] {
     top: 39px;
   }
+
   #command-bar {
     position: fixed;
     top: 60px;
@@ -132,56 +158,67 @@
     border-left-color: rgba(229, 229, 229, 0.596);
     border-left-width: 1px;
   }
+
   #command-bar[small] {
     top: 39px;
   }
+
   #command-bar > .wrapper {
     position: relative;
     left: 50%;
     transform: translateX(-50%);
   }
+
   #command-bar > .wrapper .command {
     margin: 5px 0;
   }
+
   .auth-logo {
     width: auto;
     height: 22px;
     margin-left: -5px;
     margin-top: 6.5px;
   }
+
   .auth-logo[small] {
     width: auto;
     height: 20px;
     margin-left: -5px;
     margin-top: 4.7px;
   }
+
   .auth-company {
     font-size: 26px;
     font-weight: bold;
     margin: 0;
   }
+
   .auth-company[small] {
     font-size: 23px;
     font-weight: bold;
     margin: 0;
   }
+
   .copyright-info {
     text-align: center;
     color: #333333;
     font-size: 12px;
     margin: 10px 0;
   }
+
   .hero {
     position: absolute;
     top: 50%;
     left: 50%;
     transform: translate(-50%, -50%);
   }
+
   .hero-x {
     position: absolute;
     left: 50%;
     transform: translateX(-50%);
   }
+
   .page-hero {
     position: fixed;
     top: 50%;
@@ -189,6 +226,7 @@
     transform: translate(-50%, -50%);
     z-index: 600;
   }
+
   .page-overlay {
     position: fixed;
     top: 0;
@@ -197,7 +235,9 @@
     bottom: 0;
     background: rgba(0, 0, 0, 0.623);
   }
+
   .wrapper {
     padding: 10px;
   }
+
 </style>
