+++ src/renderer/initialStates/auth.js
@@ -1,0 +1,4 @@
+export default {
+  token: '',
+  domain: ''
+}
