+++ src/renderer/initialStates/config.js
@@ -1,0 +1,58 @@
+export default {
+  platform: '',
+  info: {
+    snackbar: {
+      status: false,
+      message: ''
+    }
+  },
+  loggedIn: false,
+  dialogs: {
+    about: {
+      active: false
+    },
+    properties: {
+      active: false
+    },
+    delete: {
+      active: false,
+      options: null
+    },
+    newFolder: {
+      active: false,
+      options: null
+    },
+    dataTransfer: {
+      active: false,
+      options: {
+        list: [],
+        type: -1
+      }
+    },
+    transferState: {
+      active: false,
+      list: []
+    },
+    alert: {
+      active: false,
+      options: {
+        title: '',
+        message: ''
+      }
+    },
+    settings: {
+      active: false,
+      options: {}
+    },
+    disconnect: {
+      active: false,
+      options: {}
+    },
+    persistantLoader: {
+      active: false,
+      options: {}
+    }
+  },
+  drag: false,
+  settings: []
+}
