+++ src/renderer/initialStates/nav.js
@@ -1,0 +1,14 @@
+export default {
+  rootFs: [],
+  fetchWait: false,
+  selection: [],
+  folderEmpty: {
+    valid: false,
+    path: ''
+  },
+  selectedItem: '',
+  populateSelection: {},
+  prevPath: '',
+  currentPath: '',
+  navStack: []
+}
