+++ src/renderer/store/modules/config/actions.js
@@ -1,17 +1,27 @@
+/**
+ * Actions for application User interface configuration.
+ */
+
 import { ipcRenderer } from 'electron'
 import { platform } from 'os'
 import appConfig from '@/app.config.js'
+
 const { createLogger, format, transports } = require('winston')
 const { combine, timestamp, label, prettyPrint } = format
-
 const fsExtra = require('fs-extra')
 const nodePath = require('path')
 const uniqid = require('uniqid')
 const remote = require('electron').remote
 
-let logger = null
+// Timer used for drag and drop banner
 let dragOverTimer = null
 
+// Handle for Application logs
+let logger = null
+
+/**
+ * Called when SQL data is ready
+ */
 function registerForSqlReady (context) {
   ipcRenderer.on('sql_ready', function (event, data) {
     if (data.error) {
@@ -26,12 +36,24 @@ function registerForSqlReady (context) {
 }
 
 export default {
+  /**
+   * Write a log with level (error, info, warning) and message
+   */
+  writeLog: function (context, { level, message }) {
+    if (level && typeof message === 'string') {
+      logger.log({ level, message })
+    }
+  },
+
+  /**
+   * Initialization code when when application launched
+   */
   init: function (context) {
     // Set current running platform
     context.commit('setPlatform', platform())
     registerForSqlReady(context)
     ipcRenderer.send('sql', {
-      commit: 'setUser',
+      commit: 'setCredentialsFromSql',
       name: 'readFullTable',
       data: {
         table: 'user'
@@ -83,11 +105,10 @@ export default {
       })
     }
   },
-  writeLog: function (context, { level, message }) {
-    if (level && typeof message === 'string') {
-      logger.log({ level, message })
-    }
-  },
+
+  /**
+   * Read settings from database
+   */
   fetchSettings: function (context) {
     ipcRenderer.send('sql', {
       commit: 'setSettings',
@@ -97,6 +118,10 @@ export default {
       }
     })
   },
+
+  /**
+   * Write settings to database
+   */
   applySettings: function (context, data) {
     if (data && data.constructor === [].constructor) {
       if (data.length > 0) {
@@ -112,6 +137,10 @@ export default {
       }
     }
   },
+
+  /**
+   * Write new entry into SQL database
+   */
   writeSql: function (context, data) {
     ipcRenderer.send('sql', {
       commit: null,
@@ -119,6 +148,10 @@ export default {
       data: data
     })
   },
+
+  /**
+   * Update existing settings in SQL database
+   */
   updateSettings: function (context, { key, value }) {
     if (key) {
       context.dispatch('writeSql', {
@@ -129,7 +162,11 @@ export default {
       context.commit('setSettingsEntry', { key, value })
     }
   },
-  getSetting: function (context, { key }) {
+
+  /**
+   * Get specified settings from SQL database
+   */
+  /* getSetting: function (context, { key }) {
     return new Promise((resolve, reject) => {
       const settings = context.getters.getSettings
       if (settings && settings.constructor === [].constructor) {
@@ -156,16 +193,32 @@ export default {
         }
       }
     })
-  },
+  }, */
+
+  /**
+   * Show snack bar notification
+   */
   showInfoSnackbar: function (context, data) {
     context.commit('setInfoSnackbar', data)
   },
+
+  /**
+   * Set authentication state of the application
+   */
   authState: function (context, status) {
     context.commit('setAuthState', status)
   },
+
+  /**
+   * Open a specified dialog with options
+   */
   openDialog: function (context, { name, options }) {
     context.commit('setActiveDialog', { name, options })
   },
+
+  /**
+   * Toggle a specified dialog (show/hide) based on current state
+   */
   toggleDialog: function (context, { name, options }) {
     const dialogs = context.getters.getDialogs
     const active = dialogs && dialogs[`${name}`] && dialogs[`${name}`].active
@@ -175,9 +228,17 @@ export default {
       context.commit('setActiveDialog', { name, options })
     }
   },
+
+  /**
+   * Close a specified dialog
+   */
   closeDialog: function (context, { name }) {
     context.commit('setInertDialog', { name })
   },
+
+  /**
+   * Show drag banner (or) info graphics
+   */
   showDrag: function (context, e) {
     e && e.preventDefault()
     if (dragOverTimer) {
@@ -189,10 +250,18 @@ export default {
       context.dispatch('hideDrag')
     }, 200)
   },
+
+  /**
+   * Hide drag banner (or) info graphics
+   */
   hideDrag: function (context, e) {
     e && e.preventDefault()
     context.commit('setDragInert')
   },
+
+  /**
+   * Handler when a file dropped using drag and drop on to the application
+   */
   dropFile: function (context, event) {
     const pwd = context.getters.getCurrentPath
     const files = Object.assign([], event.dataTransfer.files)
@@ -212,6 +281,10 @@ export default {
       })
     }
   },
+
+  /**
+   * Toggle pre-selection Upload/Download transfer state of recently dropped file
+   */
   toggleListDataTransfer: function (context, { id }) {
     if (id) {
       const list = context.getters.getListDataTransfer
@@ -225,12 +298,25 @@ export default {
       }
     }
   },
+
+  /**
+   * Add a Upload/Download job to the wait queue
+   */
   waitListJob: function (context, data) {
     context.commit('setWaitListJob', data)
   },
+
+  /**
+   * Start the job queue
+   */
   startListJob: function (context, data) {
     context.commit('setStartWaitListJob', data)
   },
+
+  /**
+   * Handler for successful data transfer.
+   * This will set the job transfer (upload/download) as success.
+   */
   doneTransfer: function (context, data) {
     context.commit('setDoneTransfer', data)
     context.dispatch('fetchSelection', {
@@ -238,22 +324,39 @@ export default {
       is_dir: true
     })
   },
+
+  /**
+   * Handler for abort data transfer.
+   * This will set the job transfer (upload/download) as abort.
+   */
   abortTransfer: function (context, data) {
     context.commit('setAbortTransfer', { transferId: data.transferId })
     // context.commit('setAbortTransfer', data)
   },
-  updateJobProgress: function (context, data) {
-    context.commit('setJobProgress', data)
-  },
-  clearConfigStates: function (context) {
-    context.commit('resetConfigStates')
-  },
+
+  /**
+   * Clear list with transfer state done or abort
+   */
   clearFinished: function (context) {
-    const List = Object.assign([], context.getters.getTransferStates)
+    const List = Object.assign([], context.getters.getTransferStateList)
     List.forEach((listItem) => {
       if (listItem && (listItem.done || listItem.abort)) {
         context.commit('clearTransferListItem', { transferId: listItem.transferId })
       }
     })
+  },
+
+  /**
+   * Update a specified job progress
+   */
+  updateJobProgress: function (context, data) {
+    context.commit('setJobProgress', data)
+  },
+
+  /**
+   * CleanUP:: Configuration data on logout
+   */
+  clearConfigStates: function (context) {
+    context.commit('resetConfigStates')
   }
 }
