+++ src/renderer/store/modules/navigator/actions.js
@@ -1,14 +1,23 @@
+/**
+ * Actions for application file navigator/explorer.
+ */
+
 import axios from 'axios'
 import appConfig from '@/app.config.js'
 import TransferActivity from '@/threads/TransferActivity'
-
-let transferActivity = null
 
 const fs = require('fs')
 const nodePath = require('path')
 const base64 = require('base64-js')
 
+// Transfer Activity Handle
+let transferActivity = null
+
 export default {
+  /**
+   * Initialize Transfer Activity.
+   * Transfer Activity handles upload and download operations
+   */
   initTransferActivity: function (context, { threadCount }) {
     if (transferActivity === null) {
       transferActivity = new TransferActivity({
@@ -41,6 +50,11 @@ export default {
       })
     }
   },
+
+  /**
+   * Get status of DBFS remote server.
+   * It is used to verify the given domain and bearer token.
+   */
   getStatus: function (context, data) {
     const url = context.getters.getDomain
     const token = context.getters.getToken
@@ -53,6 +67,10 @@ export default {
       }
     )
   },
+
+  /**
+   * Update application root folder.
+   */
   updateRootFs: function (context, data) {
     if (
       data && data.constructor === [].constructor
@@ -60,6 +78,10 @@ export default {
       context.commit('setRootFs', data)
     }
   },
+
+  /**
+   * Get list of files and folders of root folder
+   */
   fetchRootFs: function (context) {
     const url = context.getters.getDomain
     const token = context.getters.getToken
@@ -76,9 +98,17 @@ export default {
       }
     })
   },
+
+  /**
+   * Clear contents files/folders inside selected folder.
+   */
   clearSelection: function (context) {
     context.commit('setSelectionEmpty')
   },
+
+  /**
+   * Get list of files and folders of a selected folder
+   */
   fetchSelection: function (context, { path }) {
     const url = context.getters.getDomain
     const token = context.getters.getToken
@@ -117,6 +147,10 @@ export default {
       })
     })
   },
+
+  /**
+   * Create a new folder at specified path
+   */
   createNewFolder: function (context, { path, folderName }) {
     const url = context.getters.getDomain
     const token = context.getters.getToken
@@ -142,14 +176,27 @@ export default {
       }
     })
   },
+
+  /**
+   * Highlight UI selection a folder or file.
+   * This will highlight select file or folder with background color.
+   */
   selectItem: function (context, { path }) {
     if (path) {
-      context.commit('setSelectedItem', path)
+      context.commit('setSelectedPath', path)
     }
   },
+
+  /**
+   * Clear UI selection of a folder or file.
+   */
   clearItem: function (context) {
-    context.commit('clearSelectedItem')
+    context.commit('clearSelectedPath')
   },
+
+  /**
+   * Delete selected remote file or folder.
+   */
   deleteSelected: function (context, { path, pwd }) {
     const url = context.getters.getDomain
     const token = context.getters.getToken
@@ -177,11 +224,19 @@ export default {
       }
     })
   },
+
+  /**
+   * Add new transfer job (upload/download)
+   */
   addJob: async function (context, data) {
     if (transferActivity) {
       transferActivity.addJob(data)
     }
   },
+
+  /**
+   * Prepare upload of specified file.
+   */
   prepareUpload: function (context, { options }) {
     // Get target working directory path where file will be created
     const targetPath = context.getters.getCurrentPath
@@ -243,6 +298,10 @@ export default {
       })
     }
   },
+
+  /**
+   * Prepare upload of specified file.
+   */
   prepareDownload: function (context, { options }) {
     if (options.list.length > 0) {
       const url = context.getters.getDomain
@@ -263,9 +322,17 @@ export default {
       })
     }
   },
+
+  /**
+   * Cancel upload or download of specified file.
+   */
   cancelTransfer: function (context, data) {
     transferActivity.cancelJob(data)
   },
+
+  /**
+   * Cancel upload or download of all files in transfer list.
+   */
   cancelAllTransfers: function (context, data) {
     return new Promise((resolve, reject) => {
       const list = context.getters.getTransferStateList
@@ -275,9 +342,18 @@ export default {
       resolve()
     })
   },
+
+  /**
+   * Set previous path in action bar
+   */
   setPrevPath: function (context, { path }) {
     context.commit('setPrevPath', path)
   },
+
+  /**
+   * Push/Add a folder name to navigation stack.
+   * Navigation stacks used to generate relative paths of selected folder.
+   */
   pushNavStack: function (context, { path }) {
     const item = path && path.split('/')[path.split('/').length - 1]
     item && context.commit(
@@ -286,13 +362,28 @@ export default {
     )
     !item && context.dispatch('clearNavStack')
   },
+
+  /**
+   * Pop/Remove a folder name from navigation stack.
+   * Navigation stacks used to generate relative paths of selected folder.
+   */
   popNavStack: function (context) {
     context.commit('popNavStack')
   },
+
+  /**
+   * Clear navigation stack by replacing with, '/', root.
+   * Navigation stacks used to generate relative paths of selected folder.
+   */
   clearNavStack: function (context) {
     context.commit('clearNavStack')
   },
-  clearNavigatorStates: function (context) {
+
+  /**
+   * Clear navigation initial states.
+   * clearNavigatorInitialStates will be called when user logout.
+   */
+  clearNavigatorInitialStates: function (context) {
     context.commit('resetNavigatorStates')
   }
 }
