+++ src/renderer/store/modules/navigator/getters.js
@@ -1,16 +1,18 @@
+/**
+ * Getters for application file navigator/explorer.
+ */
+
 export default {
+  /**
+   * Get (Array) root folder.
+   */
   getRootFs: function (state) {
     return state.rootFs
   },
-  getSelection: function (state) {
-    return state.selection
-  },
-  getSelectedItem: function (state) {
-    return state.selectedItem
-  },
-  getPrevPath: function (state) {
-    return state.prevPath
-  },
+
+  /**
+   * Get (String) File navigator current working path.
+   */
   getCurrentPath: function (state) {
     return `/${state.navStack.join('/')}`
   }
